+++ Boa.Constrictor.Playwright/CHANGELOG.md
@@ -13,7 +13,14 @@ Its format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Added `PlaywrightLocator` to store description + selector logic of a playwright locator
 - Added `AbstractPageTask` that makes it easier to write Tasks that use the BrowseTheWebSynchronously Ability
 - Added `AbstractLocatorTask` to make it easier to write tasks that perform operations on a single locator
-- Added `Click`, `Fill`, `SetChecked` and `Go` Tasks
+- Added the following Tasks:
+  - `Click`
+  - `DblClick`
+  - `Fill`
+  - `SetChecked`
+  - `SelectOption`
+  - `Go`
+  - `Hover`
 - Added `OpenNewPage` task to initialize page and navigate to the specified url
 - Added `Text` and `Attribute` questions
 - Added an `Expects` extension method to `IActor` to make it easier to use playwrights built in `ILocatorAssertions`